requirements.txt
Babel==2.4.0
certifi==2017.4.17
chardet==3.0.3
Django==1.11.2
django-phonenumber-field==1.3.0
djangorestframework==3.6.3
httpie==0.9.9
idna==2.5
mysqlclient==1.3.10
phonenumberslite==8.5.0
Pygments==2.2.0
pytz==2017.2
requests==2.17.3
urllib3==1.21.1
